backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
mdurl==0.1.2
psycopg2-binary==2.9.9
Pygments==2.16.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
